# internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Blockyard is a platform for deploying and managing Shiny applications.",
        "title": "Blockyard API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/tasks/{taskID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the current status (running, completed, failed) of a background task.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Get task status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "taskID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.taskStatusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{taskID}/logs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Stream log output for a background task (e.g. bundle restore). Returns buffered output, then follows live lines until task completes.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Stream task logs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "taskID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Log output (text/plain, chunked)"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.taskStatusResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Personal access token (PAT). Format: \"Bearer by_...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}